myAPIJWT/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
python-multipart 